+++ lib/messages_module.rb
@@ -1,0 +1,23 @@
+
+module Messages
+
+
+
+def loading_bay_is_too_full_to_land_planes
+  "Safety hazard: The loading bay has too many planes to land"
+end
+
+def bad_weather_message
+  "Safety hazard: can't take off right now due to stormy weather"
+end
+
+
+def plane_landed_message
+  "plane has now landed at the airport\n"
+end
+
+def plane_taken_off_message
+  "The plane has now taken to the air\n"
+end
+
+end
