+++ lib/plane.rb
@@ -1,7 +1,9 @@
 require_relative 'airport'
 require_relative 'weather'
+require_relative 'messages_module'
 
 class Plane
+include Messages
 attr_accessor :flying
 
   def land_at(airport)
@@ -52,25 +54,12 @@ attr_accessor :flying
     airport.landing_bay.length == airport.capacity
   end
 
-  def loading_bay_is_too_full_to_land_planes
-    "Safety hazard: The loading bay has too many planes to land"
-  end
-
-  def bad_weather_message
-    "Safety hazard: can't take off right now due to stormy weather"
-  end
 
   def make_plane_land_in_landing_bay(airport)
     airport.landing_bay << self
   end
 
-  def plane_landed_message
-    "plane has now landed at the airport\n"
-  end
 
-  def plane_taken_off_message
-    "The plane has now taken to the air\n"
-  end
 
   def switch_to_opposite_value
     if @flying == false
